feat: setup the folder structure
